python-back/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pillow==10.2.0
numpy==1.26.3
ultralytics==8.1.0
requests==2.31.0
python-dotenv==1.0.0
cloudinary==1.37.0
# RAG Chatbot dependencies
langchain
langchain-community
langchain-groq
faiss-cpu
pypdf
sentence-transformers